+++ src/platform/macos/mod.rs
@@ -5,8 +5,8 @@ use proc_pid::ProcPid;
 use proto_listener::ProtoListener;
 use socket_fd::SocketFd;
 
-use crate::Listener;
 use crate::platform::macos::proc_path::ProcPath;
+use crate::{Listener, Process, Protocol};
 
 mod c_libproc;
 mod c_proc_fd_info;
@@ -22,7 +22,7 @@ mod statics;
 pub(crate) fn get_all() -> crate::Result<HashSet<Listener>> {
     let mut listeners = HashSet::new();
 
-    for pid in ProcPid::get_all().into_iter().flatten() {
+    for pid in ProcPid::get_all()? {
         for fd in SocketFd::get_all_of_pid(pid).iter().flatten() {
             if let Ok(proto_listener) = ProtoListener::from_pid_fd(pid, fd)
                 && let Ok(ProcName(name)) = ProcName::from_pid(pid)
@@ -46,19 +46,22 @@ pub(crate) fn get_all() -> crate::Result<HashSet<Listener>> {
     Ok(listeners)
 }
 
-// #[cfg(test)]
-// mod tests {
-//     #[test]
-//     fn test_get_all() {
-//         let listeners = crate::get_all().unwrap();
-//         assert!(!listeners.is_empty());
-//
-//         // let out = std::process::Command::new("netstat")
-//         //     .args(["-p", "tcp", "-van"])
-//         //     .output()
-//         //     .unwrap();
-//         // for l in String::from_utf8(out.stdout).unwrap().lines().filter(|l| l.contains("LISTEN")) {
-//         //     println!("{}", l);
-//         // }
-//     }
-// }
+pub(crate) fn get_process_by_port(port: u16, protocol: Protocol) -> crate::Result<Process> {
+    for pid in ProcPid::get_all()? {
+        for fd in SocketFd::get_all_of_pid(pid).iter().flatten() {
+            if let Ok(proto_listener) = ProtoListener::from_pid_fd(pid, fd)
+                && proto_listener.socket_addr().port() == port
+                && proto_listener.protocol() == protocol
+                && let Ok(ProcName(name)) = ProcName::from_pid(pid)
+            {
+                let ProcPath(path) = ProcPath::from_pid(pid);
+                let Ok(pid_u_32) = pid.as_u_32() else {
+                    continue;
+                };
+                return Ok(Process::new(pid_u_32, name, path));
+            }
+        }
+    }
+
+    Err("No process found listening on the specified port and protocol".into())
+}