+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All releases with the relative changes are documented in this file.
 
 ## [UNRELEASED]
 ### Added
+- New API `get_process_by_port` to efficiently get the process listening on a specific port and protocol ([#34](https://github.com/GyulyVGC/listeners/pull/34) — fixes [#25](https://github.com/GyulyVGC/listeners/issues/25))
 - `IS_OS_SUPPORTED` constant to check if the consumer platform is supported by the library
 - Benchmarks to measure performance on all supported platforms ([#31](https://github.com/GyulyVGC/listeners/pull/31))
+### Changed
+- Only open one `ProcFd` at a time on Linux ([#30](https://github.com/GyulyVGC/listeners/pull/30))
+### Removed
+- `get_ports_by_pid`, `get_ports_by_process_name`, and `get_processes_by_port`, since they were not optimized because internally simply based on `get_all`
 
 ## [0.3.0] - 2025-08-02
 ### Added